+++ codegen/src/codegen_go.rs
@@ -30,7 +30,8 @@ use crate::{
     gen::CodeGen,
     model::{
         get_operation, get_sorted_fields, get_trait, is_opt_namespace, value_to_json,
-        wasmcloud_core_namespace, wasmcloud_model_namespace, CommentKind, PackageName, Ty,
+        wasmcloud_actor_namespace, wasmcloud_core_namespace, wasmcloud_model_namespace,
+        CommentKind, PackageName, Ty,
     },
     render::Renderer,
     writer::Writer,
@@ -83,6 +84,7 @@ impl fmt::Display for DecodeRef {
     }
 }
 
+#[derive(Default)]
 #[allow(dead_code)]
 pub struct GoCodeGen<'model> {
     /// if set, limits declaration output to this namespace only
@@ -345,13 +347,15 @@ impl<'model> CodeGen for GoCodeGen<'model> {
             w,
             r#"package {}
             import (
-                {}
                 msgpack "github.com/wasmcloud/tinygo-msgpack" //nolint
                 cbor "github.com/wasmcloud/tinygo-cbor" //nolint
+                {}
             )"#,
             &self.package,
-            if ns != wasmcloud_model_namespace() && ns != wasmcloud_core_namespace() {
-                "\"github.com/wasmcloud/actor-tinygo\" //nolint"
+            if ns == wasmcloud_actor_namespace() {
+                "core \"github.com/wasmcloud/interfaces/core/tinygo\" //nolint"
+            } else if ns != wasmcloud_model_namespace() && ns != wasmcloud_core_namespace() {
+                "actor \"github.com/wasmcloud/actor-tinygo\" //nolint"
             } else {
                 // avoid circular dependencies - core and model are part of the actor-tinygo package
                 ""
@@ -784,9 +788,9 @@ impl<'model> GoCodeGen<'model> {
         let (fields, _is_numbered) = get_sorted_fields(ident, strukt)?;
         for member in fields.iter() {
             self.apply_documentation_traits(w, member.id(), member.traits());
-            let (field_name, _ser_name) = self.get_field_name_and_ser_name(member)?;
+            let (field_name, ser_name) = self.get_field_name_and_ser_name(member)?;
             let target = member.target();
-            let field_tags = "";
+            let field_tags = format!(r#"`json:"{}"`"#, ser_name);
             writeln!(
                 w,
                 "  {} {} {}",