+++ codegen/src/loader.rs
@@ -230,16 +230,14 @@ fn urls_to_cached_files(urls: Vec<String>) -> Result<Vec<PathBuf>> {
                                 Error::Other(format!("internal download error {}", e))
                             })?;
                             let cache_file = weld_cache.join(rel_path);
-                            std::fs::create_dir_all(&cache_file.parent().unwrap()).map_err(
-                                |e| {
-                                    Error::Io(format!(
-                                        "creating folder {}: {}",
-                                        &cache_file.parent().unwrap().display(),
-                                        e
-                                    ))
-                                },
-                            )?;
-                            std::fs::copy(&downloaded_file, &cache_file).map_err(|e| {
+                            std::fs::create_dir_all(cache_file.parent().unwrap()).map_err(|e| {
+                                Error::Io(format!(
+                                    "creating folder {}: {}",
+                                    &cache_file.parent().unwrap().display(),
+                                    e
+                                ))
+                            })?;
+                            std::fs::copy(downloaded_file, &cache_file).map_err(|e| {
                                 Error::Other(format!(
                                     "writing cache file {}: {}",
                                     &cache_file.display(),