+++ rpc-rs/src/cbor.rs
@@ -192,13 +192,10 @@ pub struct Encoder<W: Write> {
 }
 
 pub fn vec_encoder(header: bool) -> Encoder<Vec<u8>> {
-    let buf = if header {
-        let mut buf = Vec::new();
+    let mut buf = Vec::new();
+    if header {
         crate::common::MessageFormat::Cbor.write_header(&mut buf).unwrap();
-        buf
-    } else {
-        Vec::new()
-    };
+    }
     Encoder { inner: minicbor::Encoder::new(buf) }
 }