+++ rpc-rs/src/chunkify.rs
@@ -19,26 +19,25 @@
 
 use std::{
     collections::HashMap,
-    io::Read,
+    marker::Unpin,
     sync::{Arc, RwLock},
 };
 
-use nats::{
-    jetstream::JetStream,
+use async_nats::jetstream::{
+    self,
     object_store::{Config, ObjectStore},
-    JetStreamOptions,
+    Context,
 };
+use futures::TryFutureExt;
 use once_cell::sync::OnceCell;
+use tokio::io::{AsyncRead, AsyncReadExt};
 use tracing::{debug, error, instrument};
 
-use crate::{
-    error::{RpcError, RpcResult},
-    provider_main::get_host_bridge,
-};
+use crate::error::{RpcError, RpcResult};
 
 /// Maximum size of a message payload before it will be chunked
 /// Nats currently uses 128kb chunk size so this should be at least 128KB
-const CHUNK_THRESHOLD_BYTES: usize = 1024 * 700; // 700KB
+const CHUNK_THRESHOLD_BYTES: usize = 1024 * 900; // 900KB
 
 /// check if message payload needs to be chunked
 pub(crate) fn needs_chunking(payload_size: usize) -> bool {
@@ -46,7 +45,7 @@ pub(crate) fn needs_chunking(payload_size: usize) -> bool {
 }
 
 /// map from lattice to ObjectStore - includes nats client connection
-type JsMap = HashMap<String, JetStream>;
+type JsMap = HashMap<String, Context>;
 
 fn jetstream_map() -> Arc<RwLock<JsMap>> {
     static INSTANCE: OnceCell<Arc<RwLock<JsMap>>> = OnceCell::new();
@@ -64,21 +63,40 @@ pub(crate) fn shutdown() {
 #[derive(Clone)]
 pub struct ChunkEndpoint {
     lattice: String,
-    js: JetStream,
+    js: Context,
 }
 
 impl ChunkEndpoint {
-    pub fn new(lattice: String, js: JetStream) -> Self {
+    pub fn new(lattice: String, js: Context) -> Self {
         ChunkEndpoint { lattice, js }
+    }
+
+    pub(crate) fn with_client(
+        lattice: String,
+        nc: async_nats::Client,
+        domain: Option<String>,
+    ) -> Self {
+        let map = jetstream_map();
+        let mut _w = map.write().unwrap(); // panics if lock is poisoned
+        let js = _w.get(&lattice).cloned().unwrap_or_else(|| {
+            let js = if let Some(domain) = domain {
+                jetstream::with_domain(nc, domain)
+            } else {
+                jetstream::new(nc)
+            };
+            _w.insert(lattice.clone(), js.clone());
+            js
+        });
+        ChunkEndpoint::new(lattice, js)
     }
 
     /// load the message after de-chunking
     #[instrument(level = "trace", skip(self))]
-    pub fn get_unchunkified(&self, inv_id: &str) -> RpcResult<Vec<u8>> {
+    pub async fn get_unchunkified(&self, inv_id: &str) -> RpcResult<Vec<u8>> {
         let mut result = Vec::new();
-        let store = self.create_or_reuse_store()?;
+        let store = self.create_or_reuse_store().await?;
         debug!(invocation_id = %inv_id, "chunkify starting to receive");
-        let mut obj = store.get(inv_id).map_err(|e| {
+        let mut obj = store.get(inv_id).await.map_err(|e| {
             RpcError::Nats(format!(
                 "error starting to receive chunked stream for inv {}:{}",
                 inv_id, e
@@ -90,8 +108,9 @@ impl ChunkEndpoint {
                 "error receiving chunked stream for inv {}:{}",
                 inv_id, e
             ))
-        })?;
-        if let Err(e) = store.delete(inv_id) {
+        })
+        .await?;
+        if let Err(e) = store.delete(inv_id).await {
             // not deleting will be a non-fatal error for the receiver,
             // if all the bytes have been received
             error!(invocation_id = %inv_id, error = %e, "deleting chunks for inv");
@@ -100,18 +119,23 @@ impl ChunkEndpoint {
     }
 
     /// load response after de-chunking
-    pub fn get_unchunkified_response(&self, inv_id: &str) -> RpcResult<Vec<u8>> {
+    pub async fn get_unchunkified_response(&self, inv_id: &str) -> RpcResult<Vec<u8>> {
         // responses are stored in the object store with '-r' suffix on the object name
-        self.get_unchunkified(&format!("{}-r", inv_id))
+        self.get_unchunkified(&format!("{}-r", inv_id)).await
     }
 
     /// chunkify a message
     #[instrument(level = "trace", skip(self, bytes))]
-    pub fn chunkify(&self, inv_id: &str, bytes: &mut impl Read) -> RpcResult<()> {
-        let store = self.create_or_reuse_store()?;
+    pub async fn chunkify(
+        &self,
+        inv_id: &str,
+        mut bytes: (impl AsyncRead + Unpin),
+    ) -> RpcResult<()> {
+        let store = self.create_or_reuse_store().await?;
         debug!(invocation_id = %inv_id, "chunkify starting to send");
         let info = store
-            .put(inv_id, bytes)
+            .put(inv_id, &mut bytes)
+            .await
             .map_err(|e| RpcError::Nats(format!("writing chunkified for {}: {}", inv_id, e)))?;
         debug!(?info, invocation_id = %inv_id, "chunkify completed writing");
 
@@ -119,47 +143,26 @@ impl ChunkEndpoint {
     }
 
     /// chunkify a portion of a response
-    pub fn chunkify_response(&self, inv_id: &str, bytes: &mut impl Read) -> Result<(), RpcError> {
-        self.chunkify(&format!("{}-r", inv_id), bytes)
+    pub async fn chunkify_response(
+        &self,
+        inv_id: &str,
+        bytes: &mut (impl AsyncRead + Unpin),
+    ) -> Result<(), RpcError> {
+        self.chunkify(&format!("{}-r", inv_id), bytes).await
     }
 
-    fn create_or_reuse_store(&self) -> RpcResult<ObjectStore> {
-        let store = match self.js.object_store(&self.lattice) {
+    async fn create_or_reuse_store(&self) -> RpcResult<ObjectStore> {
+        let store = match self.js.get_object_store(&self.lattice).await {
             Ok(store) => store,
             Err(_) => self
                 .js
-                .create_object_store(&Config {
+                .create_object_store(Config {
                     bucket: self.lattice.clone(),
                     ..Default::default()
                 })
+                .await
                 .map_err(|e| RpcError::Nats(format!("Failed to create store: {}", &e)))?,
         };
         Ok(store)
     }
-}
-
-pub(crate) fn chunkify_endpoint(
-    domain: Option<String>,
-    lattice: String,
-) -> RpcResult<ChunkEndpoint> {
-    let js = connect_js(domain, &lattice)?;
-    Ok(ChunkEndpoint::new(lattice, js))
-}
-
-pub(crate) fn connect_js(domain: Option<String>, lattice_prefix: &str) -> RpcResult<JetStream> {
-    let map = jetstream_map();
-    let mut _w = map.write().unwrap(); // panics if lock is poisioned
-    let js: JetStream = if let Some(js) = _w.get(lattice_prefix) {
-        js.clone()
-    } else {
-        let nc = get_host_bridge().new_sync_client()?;
-        let mut jsoptions = JetStreamOptions::new();
-        if let Some(domain) = domain {
-            jsoptions = jsoptions.domain(domain.as_str());
-        }
-        let js = JetStream::new(nc, jsoptions);
-        _w.insert(lattice_prefix.to_string(), js.clone());
-        js
-    };
-    Ok(js)
 }