+++ rpc-rs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # wasmbus-rpc Changelog
 
+## 0.11.0-alpha.1
+
+- upgrade to async-nats 0.22.0 
+- additional logging on reconnect 
+- add json tags to generated tinygo struct declarations
+- bumped chunkify threshold to 900KiB \(900\*1024\) to match host threshold
+ 
+
 ## 0.10.0
 
 - upgrade to async-nats 0.18.0